requirements-dev.txt
# testing requirements
testing.postgresql~=1.3.0
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
pytest-env==1.2.0
deepdiff==8.6.1
blinker==1.9.0
requests_mock==1.12.1
respx~=0.22.0

# GRPC
grpcio-testing==1.78.0